more parse iso8601 to module
